+++ CHANGELOG.md
@@ -14,7 +14,7 @@
 * [#498 Add preferences to shows export](https://github.com/WebTools-NG/WebTools-NG/issues/498)
 * [#500 Library Path Mapping](https://github.com/WebTools-NG/WebTools-NG/issues/500)
 * [#505 FindMedia](https://github.com/WebTools-NG/WebTools-NG/issues/505)
-
+* [#509 Refactor Settings](https://github.com/WebTools-NG/WebTools-NG/issues/509)
 
 ## V0.3.17 (20220601)